Fix bug with identity and move some stuff around
