+++ src/storage_broker/utils/broker_logging.py
@@ -7,7 +7,7 @@ import watchtower
 from logstash_formatter import LogstashFormatterV1
 from boto3.session import Session
 
-from ..utils import config
+from storage_broker.utils import config
 
 
 def config_cloudwatch(logger):
